+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ### is.codion.common.model
 - Unused slf4j dependency removed.
 - ConditionModel.identifier() removed, related refactoring.
+- ConditionModel.AutomaticWildcard renamed Wildcard.
 ### is.codion.swing.common.model
 - DefaultFilterTableModel, null item rejection improved.
 - DefaultFilterTableSelection, null item rejection improved.